refactor KubeVirt images provisioning
